+++ packages/react/src/AsyncBoundary.tsx
@@ -1,9 +1,8 @@
-import type { ComponentProps, ComponentRef, ComponentType, SuspenseProps } from 'react'
+import type { ComponentRef, SuspenseProps } from 'react'
 import { forwardRef } from 'react'
 import { ErrorBoundary } from './ErrorBoundary'
 import type { ErrorBoundaryProps } from './ErrorBoundary'
 import { Suspense } from './Suspense'
-import type { PropsWithoutChildren } from './utility-types'
 
 /**
  * @deprecated Use SuspenseProps and ErrorBoundaryProps as alternatives
@@ -44,48 +43,3 @@ export const AsyncBoundary = Object.assign(BaseAsyncBoundary, {
    */
   CSROnly,
 })
-
-/**
- * @deprecated Use wrap.ErrorBoundary().Suspense().on as alternatives
- */
-export const withAsyncBoundary = Object.assign(
-  <TProps extends ComponentProps<ComponentType> = Record<string, never>>(
-    Component: ComponentType<TProps>,
-    asyncBoundaryProps: PropsWithoutChildren<AsyncBoundaryProps>
-  ) => {
-    const Wrapped = (props: TProps) => (
-      <AsyncBoundary {...asyncBoundaryProps}>
-        <Component {...props} />
-      </AsyncBoundary>
-    )
-
-    if (process.env.NODE_ENV !== 'production') {
-      const name = Component.displayName || Component.name || 'Component'
-      Wrapped.displayName = `withAsyncBoundary(${name})`
-    }
-
-    return Wrapped
-  },
-  {
-    /**
-     * @deprecated Use wrap.ErrorBoundary().Suspense.CSROnly().on as alternatives
-     */
-    CSROnly: <TProps extends ComponentProps<ComponentType> = Record<string, never>>(
-      Component: ComponentType<TProps>,
-      asyncBoundaryProps: PropsWithoutChildren<AsyncBoundaryProps>
-    ) => {
-      const Wrapped = (props: TProps) => (
-        <AsyncBoundary.CSROnly {...asyncBoundaryProps}>
-          <Component {...props} />
-        </AsyncBoundary.CSROnly>
-      )
-
-      if (process.env.NODE_ENV !== 'production') {
-        const name = Component.displayName || Component.name || 'Component'
-        Wrapped.displayName = `withAsyncBoundary.CSROnly(${name})`
-      }
-
-      return Wrapped
-    },
-  }
-)
